+++ src/replio/config.py
@@ -22,6 +22,7 @@ DEFAULT_CONFIG = {
     'markdown_streaming': False,
     'show_context_size': True,
     'clear_screen': True,
+    'show_version': True,
     'compact_keep': 4,
     'noise_tools': ['fetch_page'],
     'web_search': False,